Plugin refactoring and cleanup

+++ plugins/bl/regionalinvestment/updates/builder_table_update_bl_regionalinvestment_investment_opportunities.php
@@ -1,4 +1,6 @@
-<?php namespace BL\RegionalInvestment\Updates;
+<?php
+
+namespace BL\RegionalInvestment\Updates;
 
 use Schema;
 use October\Rain\Database\Updates\Migration;
@@ -7,17 +9,19 @@ class BuilderTableUpdateBlRegionalinvestmentInvestmentOpportunities extends Migr
 {
     public function up()
     {
-        Schema::table('bl_regionalinvestment_investment_opportunities', function($table)
-        {
-            $table->text('photos')->nullable();
-        });
+        if (!Schema::hasColumn('bl_regionalinvestment_investment_opportunities', 'photos')) {
+            Schema::table('bl_regionalinvestment_investment_opportunities', function ($table) {
+                $table->text('photos')->nullable();
+            });
+        }
     }
-    
+
     public function down()
     {
-        Schema::table('bl_regionalinvestment_investment_opportunities', function($table)
-        {
-            $table->dropColumn('photos');
-        });
+        if (Schema::hasColumn('bl_regionalinvestment_investment_opportunities', 'photos')) {
+            Schema::table('bl_regionalinvestment_investment_opportunities', function ($table) {
+                $table->dropColumn('photos');
+            });
+        }
     }
 }
